requirements.txt
# CloudStream Desktop — Python dependencies
# Install with: pip install -r requirements.txt

# Core UI
customtkinter>=5.2.2
Pillow>=10.0.0

# HTTP / Networking
httpx[http2]>=0.27.0
requests>=2.31.0

# Cloudflare bypass (optional but recommended)
cloudscraper>=1.2.71

# HTML parsing
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Video player (optional — falls back to external mpv/vlc)
# python-mpv requires mpv to be installed on the system first
# Windows: https://mpv.io/installation/
# Linux: sudo apt install mpv python3-mpv
# After mpv is installed: pip install python-mpv
# python-mpv>=1.0.6

# Subtitle handling
pysrt>=1.1.2
# ass>=0.5.2   # for .ass/.ssa subtitles

# DEX / CS3 decompiler
androguard>=4.1.3

# Utilities
# thefuzz>=0.22.1    # fuzzy search (optional)
# python-dotenv>=1.0.0
